// src/types/PowerlineConfig.ts
import { z } from 'zod';

export const PowerlineConfigSchema = z.object({
    enabled: z.boolean().default(false),
    separators: z.array(z.string()).default(['\uE0B0']),
    separatorInvertBackground: z.array(z.boolean()).default([false]),
    startCaps: z.array(z.string()).default([]),
    endCaps: z.array(z.string()).default([]),
    theme: z.string().optional(),
    autoAlign: z.boolean().default(false),
    continueThemeAcrossLines: z.boolean().default(false)
});

export type PowerlineConfig = z.infer<typeof PowerlineConfigSchema>;